backend/pyproject.toml
[project]
name = "istockpick-backend"
version = "1.1.0"
description = "iStockPick AI-powered stock analysis API"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "pandas",
    "numpy",
    "capitolgains",
    "openbb",
    "pyqlib",
    "yfinance",
]
